+++ src/solvers/levmar/mod.rs
@@ -1,23 +1,24 @@
 use crate::prelude::*;
 use crate::statistics::FitStatistics;
+use crate::util::{to_vector, Weights};
+pub use builder::LevMarProblemBuilder;
+/// type alias for the solver of the [levenberg_marquardt](https://crates.io/crates/levenberg-marquardt) crate
+// pub use levenberg_marquardt::LevenbergMarquardt as LevMarSolver;
+use levenberg_marquardt::LevenbergMarquardt;
 use levenberg_marquardt::{LeastSquaresProblem, MinimizationReport};
 use nalgebra::storage::Owned;
 use nalgebra::{
     ComplexField, DMatrix, DefaultAllocator, Dim, DimMin, Dyn, Matrix, MatrixView, OMatrix,
     OVector, RawStorageMut, RealField, Scalar, UninitMatrix, Vector, VectorView, SVD, U1,
 };
+use num_traits::{Float, FromPrimitive};
+#[cfg(feature = "parallel")]
+use rayon::prelude::*;
+use std::ops::Mul;
 
 mod builder;
 #[cfg(any(test, doctest))]
 mod test;
-
-use crate::util::{to_vector, Weights};
-pub use builder::LevMarProblemBuilder;
-/// type alias for the solver of the [levenberg_marquardt](https://crates.io/crates/levenberg-marquardt) crate
-// pub use levenberg_marquardt::LevenbergMarquardt as LevMarSolver;
-use levenberg_marquardt::LevenbergMarquardt;
-use num_traits::{Float, FromPrimitive};
-use std::ops::Mul;
 
 /// A thin wrapper around the
 /// [`LevenbergMarquardt`](https://docs.rs/levenberg-marquardt/latest/levenberg_marquardt/struct.LevenbergMarquardt.html)
@@ -40,7 +41,7 @@ where
 {
     /// the final state of the fitting problem after the
     /// minimization finished (regardless of whether fitting was successful or not)
-    pub problem: LevMarProblem<Model, MRHS>,
+    pub problem: LevMarProblem<Model, MRHS, PARALLEL_NO>,
 
     /// the minimization report of the underlying solver.
     /// It contains information about the minimization process
@@ -118,7 +119,7 @@ where
 {
     /// internal helper for constructing an instance
     fn new(
-        problem: LevMarProblem<Model, MRHS>,
+        problem: LevMarProblem<Model, MRHS, PARALLEL_NO>,
         minimization_report: MinimizationReport<Model::ScalarType>,
     ) -> Self {
         Self {
@@ -145,40 +146,10 @@ impl<Model, const MRHS: bool> LevMarSolver<Model, MRHS>
 where
     Model: SeparableNonlinearModel,
 {
-    /// create a new solver with default parameters. Uses the underlying
-    /// solver of the `levenberg_marquardt` crate
-    #[deprecated(
-        since = "0.8.0",
-        note = "use the default() method instead. If you want to specify solver parameters use the with_solver(...) constructor."
-    )]
-    pub fn new() -> Self
-    where
-        Model::ScalarType: RealField + Float,
-    {
-        Self::default()
-    }
-
     /// creata a new solver using the given underlying solver. This allows
     /// us to configure the underlying solver with non-default parameters
     pub fn with_solver(solver: LevenbergMarquardt<Model::ScalarType>) -> Self {
         Self { solver }
-    }
-
-    /// Solve the fitting problem. This method is deprecated, use the fit method instead.
-    #[deprecated(since = "0.7.0", note = "use the fit(...) method instead")]
-    pub fn minimize(
-        &self,
-        problem: LevMarProblem<Model, MRHS>,
-    ) -> (
-        LevMarProblem<Model, MRHS>,
-        MinimizationReport<Model::ScalarType>,
-    )
-    where
-        Model: SeparableNonlinearModel,
-        LevMarProblem<Model, MRHS>: LeastSquaresProblem<Model::ScalarType, Dyn, Dyn>,
-        Model::ScalarType: Scalar + ComplexField + Copy + RealField + Float + FromPrimitive,
-    {
-        self.solver.minimize(problem)
     }
 
     /// Try to solve the given varpro minimization problem. The parameters of
@@ -194,18 +165,18 @@ where
     /// On failure (when the minimization was not deemeed successful), returns
     /// an error with the same information as in the success case.
     #[allow(clippy::result_large_err)]
-    pub fn fit(
+    pub fn fit<const PAR: bool>(
         &self,
-        problem: LevMarProblem<Model, MRHS>,
+        problem: LevMarProblem<Model, MRHS, PAR>,
     ) -> Result<FitResult<Model, MRHS>, FitResult<Model, MRHS>>
     where
         Model: SeparableNonlinearModel,
-        LevMarProblem<Model, MRHS>: LeastSquaresProblem<Model::ScalarType, Dyn, Dyn>,
+        LevMarProblem<Model, MRHS, PAR>: LeastSquaresProblem<Model::ScalarType, Dyn, Dyn>,
         Model::ScalarType: Scalar + ComplexField + RealField + Float + FromPrimitive,
     {
         #[allow(deprecated)]
-        let (problem, report) = self.minimize(problem);
-        let result = FitResult::new(problem, report);
+        let (problem, report) = self.solver.minimize(problem);
+        let result = FitResult::new(problem.into_sequential(), report);
         if result.was_successful() {
             Ok(result)
         } else {
@@ -232,13 +203,13 @@ where
     /// with a single right hand side. If this function is invoked on a problem
     /// with multiple right hand sides, an error is returned.
     #[allow(clippy::result_large_err)]
-    pub fn fit_with_statistics(
+    pub fn fit_with_statistics<const PAR: bool>(
         &self,
-        problem: LevMarProblem<Model, false>,
+        problem: LevMarProblem<Model, false, PAR>,
     ) -> Result<(FitResult<Model, false>, FitStatistics<Model>), FitResult<Model, false>>
     where
         Model: SeparableNonlinearModel,
-        LevMarProblem<Model, false>: LeastSquaresProblem<Model::ScalarType, Dyn, Dyn>,
+        LevMarProblem<Model, false, PAR>: LeastSquaresProblem<Model::ScalarType, Dyn, Dyn>,
         Model::ScalarType: Scalar + ComplexField + RealField + Float,
     {
         let FitResult {
@@ -265,6 +236,7 @@ where
         }
     }
 }
+
 impl<Model, const MRHS: bool> Default for LevMarSolver<Model, MRHS>
 where
     Model: SeparableNonlinearModel,
@@ -329,7 +301,7 @@ where
 /// coefficient vectors and data vectors respectively.
 #[derive(Clone)]
 #[allow(non_snake_case)]
-pub struct LevMarProblem<Model, const MRHS: bool>
+pub struct LevMarProblem<Model, const MRHS: bool, const PARALLEL: bool>
 where
     Model: SeparableNonlinearModel,
     Model::ScalarType: Scalar + ComplexField + Copy,
@@ -356,7 +328,55 @@ where
     cached: Option<CachedCalculations<Model::ScalarType, Dyn, Dyn>>,
 }
 
-impl<Model, const MRHS: bool> std::fmt::Debug for LevMarProblem<Model, MRHS>
+impl<Model, const MRHS: bool, const PAR: bool> LevMarProblem<Model, MRHS, PAR>
+where
+    Model: SeparableNonlinearModel,
+    Model::ScalarType: Scalar + ComplexField + Copy,
+{
+    /// convert from parallel problem to sequential one. Useful for funnelling the results
+    /// of the parallel calculations to downstream tasks that only take sequential models
+    /// for simplicity.
+    pub fn into_sequential(self) -> LevMarProblem<Model, MRHS, PARALLEL_NO> {
+        let LevMarProblem {
+            Y_w,
+            model,
+            svd_epsilon,
+            weights,
+            cached,
+        } = self;
+        LevMarProblem {
+            Y_w,
+            model,
+            svd_epsilon,
+            weights,
+            cached,
+        }
+    }
+
+    /// convert from sequential problem to a parallel one
+    pub fn into_parallel(self) -> LevMarProblem<Model, MRHS, PARALLEL_NO> {
+        let LevMarProblem {
+            Y_w,
+            model,
+            svd_epsilon,
+            weights,
+            cached,
+        } = self;
+        LevMarProblem {
+            Y_w,
+            model,
+            svd_epsilon,
+            weights,
+            cached,
+        }
+    }
+}
+
+#[allow(unused)]
+pub(crate) const PARALLEL_YES: bool = true;
+pub(crate) const PARALLEL_NO: bool = false;
+
+impl<Model, const MRHS: bool, const PAR: bool> std::fmt::Debug for LevMarProblem<Model, MRHS, PAR>
 where
     Model: SeparableNonlinearModel,
     Model::ScalarType: Scalar + ComplexField + Copy,
@@ -372,7 +392,7 @@ where
     }
 }
 
-impl<Model> LevMarProblem<Model, true>
+impl<Model, const PAR: bool> LevMarProblem<Model, true, PAR>
 where
     Model: SeparableNonlinearModel,
     Model::ScalarType: Scalar + ComplexField + Copy,
@@ -405,7 +425,7 @@ where
     }
 }
 
-impl<Model> LevMarProblem<Model, false>
+impl<Model, const PAR: bool> LevMarProblem<Model, false, PAR>
 where
     Model: SeparableNonlinearModel,
     Model::ScalarType: Scalar + ComplexField + Copy,
@@ -442,7 +462,8 @@ where
         self.Y_w.as_view()
     }
 }
-impl<Model, const MRHS: bool> LevMarProblem<Model, MRHS>
+
+impl<Model, const MRHS: bool, const PAR: bool> LevMarProblem<Model, MRHS, PAR>
 where
     Model: SeparableNonlinearModel,
     Model::ScalarType: Scalar + ComplexField + Copy,
@@ -459,7 +480,7 @@ where
 }
 
 impl<Model, const MRHS: bool> LeastSquaresProblem<Model::ScalarType, Dyn, Dyn>
-    for LevMarProblem<Model, MRHS>
+    for LevMarProblem<Model, MRHS, PARALLEL_NO>
 where
     Model::ScalarType: Scalar + ComplexField + Copy,
     <<Model as SeparableNonlinearModel>::ScalarType as ComplexField>::RealField:
@@ -566,6 +587,144 @@ where
             // performance hit in the sad path.
             let result: Result<Vec<()>, Model::Error> = jacobian_matrix
                 .column_iter_mut()
+                .enumerate()
+                .map(|(k, mut jacobian_col)| {
+                    // weighted derivative matrix
+                    let Dk = &self.weights * self.model.eval_partial_deriv(k)?; // will return none if this could not be calculated
+                    let Dk_C = Dk * linear_coefficients;
+                    let minus_ak = U * (&U_t * (&Dk_C)) - Dk_C;
+
+                    //for non-approximate jacobian we require our scalar type to be a real field (or maybe we can finagle it with clever trait bounds)
+                    //let Dk_t_rw : DVector<Model::ScalarType> = &Dk.transpose()*self.residuals().as_ref().expect("Residuals must produce result");
+                    //let _minus_bk : DVector<Model::ScalarType> = U*(&Sigma_inverse*(V_t*(&Dk_t_rw)));
+
+                    //@todo CAUTION this relies on the fact that the
+                    //elements are ordered in column major order but it avoids a copy
+                    copy_matrix_to_column(minus_ak, &mut jacobian_col);
+                    Ok(())
+                })
+                .collect::<Result<_, _>>();
+
+            // we need this check to make sure the jacobian is returned
+            // as None on error.
+            result.ok()?;
+
+            Some(jacobian_matrix)
+        } else {
+            None
+        }
+    }
+}
+
+#[cfg(feature = "parallel")]
+impl<Model, const MRHS: bool> LeastSquaresProblem<Model::ScalarType, Dyn, Dyn>
+    for LevMarProblem<Model, MRHS, PARALLEL_YES>
+where
+    Model::ScalarType: Scalar + ComplexField + Copy,
+    <<Model as SeparableNonlinearModel>::ScalarType as ComplexField>::RealField:
+        Mul<Model::ScalarType, Output = Model::ScalarType> + Float,
+    Model: SeparableNonlinearModel + std::marker::Sync,
+    DefaultAllocator: nalgebra::allocator::Allocator<Dyn>,
+{
+    type ResidualStorage = Owned<Model::ScalarType, Dyn>;
+    type JacobianStorage = Owned<Model::ScalarType, Dyn, Dyn>;
+    type ParameterStorage = Owned<Model::ScalarType, Dyn>;
+
+    #[allow(non_snake_case)]
+    /// Set the (nonlinear) model parameters `$\vec{\alpha}$` and update the internal state of the
+    /// problem accordingly. The parameters are expected in the same order that the parameter
+    /// names were provided in at model creation. So if we gave `&["tau","beta"]` as parameters at
+    /// model creation, the function expects the layout of the parameter vector to be `$\vec{\alpha}=(\tau,\beta)^T$`.
+    fn set_params(&mut self, params: &Vector<Model::ScalarType, Dyn, Self::ParameterStorage>) {
+        if self.model.set_params(params.clone()).is_err() {
+            self.cached = None;
+        }
+        // matrix of weighted model function values
+        let Phi_w = self.model.eval().ok().map(|Phi| &self.weights * Phi);
+
+        // calculate the svd
+        let svd_epsilon = self.svd_epsilon;
+        let current_svd = Phi_w.as_ref().map(|Phi_w| Phi_w.clone().svd(true, true));
+        let linear_coefficients = current_svd
+            .as_ref()
+            .and_then(|svd| svd.solve(&self.Y_w, svd_epsilon).ok());
+
+        // calculate the residuals
+        let current_residuals = Phi_w
+            .zip(linear_coefficients.as_ref())
+            .map(|(Phi_w, coeff)| &self.Y_w - &Phi_w * coeff);
+
+        // if everything was successful, update the cached calculations, otherwise set the cache to none
+        if let (Some(current_residuals), Some(current_svd), Some(linear_coefficients)) =
+            (current_residuals, current_svd, linear_coefficients)
+        {
+            self.cached = Some(CachedCalculations {
+                current_residuals,
+                current_svd,
+                linear_coefficients,
+            })
+        } else {
+            self.cached = None;
+        }
+    }
+
+    /// Retrieve the (nonlinear) model parameters as a vector `$\vec{\alpha}$`.
+    /// The order of the parameters in the vector is the same as the order of the parameter
+    /// names given on model creation. E.g. if the parameters at model creation where given as
+    /// `&["tau","beta"]`, then the returned vector is `$\vec{\alpha} = (\tau,\beta)^T$`, i.e.
+    /// the value of parameter `$\tau$` is at index `0` and the value of `$\beta$` at index `1`.
+    fn params(&self) -> Vector<Model::ScalarType, Dyn, Self::ParameterStorage> {
+        self.model.params()
+    }
+
+    /// Calculate the residual vector `$\vec{r}_w$` of *weighted* residuals at every location `$\vec{x}$`.
+    /// The residual is calculated from the data `\vec{y}` as `$\vec{r}_w(\vec{\alpha}) = W\cdot(\vec{y}-\vec{f}(\vec{x},\vec{\alpha},\vec{c}(\vec{\alpha}))$`,
+    /// where `$\vec{f}(\vec{x},\vec{\alpha},\vec{c})$` is the model function evaluated at the currently
+    /// set nonlinear parameters `$\vec{\alpha}$` and the linear coefficients `$\vec{c}(\vec{\alpha})$`. The VarPro
+    /// algorithm calculates `$\vec{c}(\vec{\alpha})$` as the coefficients that provide the best linear least squares
+    /// fit, given the current `$\vec{\alpha}$`. For more info on the math of VarPro, see
+    /// e.g. [here](https://geo-ant.github.io/blog/2020/variable-projection-part-1-fundamentals/).
+    fn residuals(&self) -> Option<Vector<Model::ScalarType, Dyn, Self::ResidualStorage>> {
+        self.cached
+            .as_ref()
+            .map(|cached| to_vector(cached.current_residuals.clone()))
+    }
+
+    #[allow(non_snake_case)]
+    /// Calculate the Jacobian matrix of the *weighted* residuals `$\vec{r}_w(\vec{\alpha})$`.
+    /// For more info on how the Jacobian is calculated in the VarPro algorithm, see
+    /// e.g. [here](https://geo-ant.github.io/blog/2020/variable-projection-part-1-fundamentals/).
+    fn jacobian(&self) -> Option<Matrix<Model::ScalarType, Dyn, Dyn, Self::JacobianStorage>> {
+        // TODO (Performance): make this more efficient by parallelizing
+        if let Some(CachedCalculations {
+            current_residuals: _,
+            current_svd,
+            linear_coefficients,
+        }) = self.cached.as_ref()
+        {
+            // this is not a great pattern, but the trait bounds on copy_from
+            // as of now prevent us from doing something more idiomatic
+            let mut jacobian_matrix = unsafe {
+                UninitMatrix::uninit(
+                    Dyn(self.model.output_len() * self.Y_w.ncols()),
+                    Dyn(self.model.parameter_count()),
+                )
+                .assume_init()
+            };
+
+            let U = current_svd.u.as_ref()?; // will return None if this was not calculated
+            let U_t = U.transpose();
+
+            //let Sigma_inverse : DMatrix<Model::ScalarType::RealField> = DMatrix::from_diagonal(&self.current_svd.singular_values.map(|val|val.powi(-1)));
+            //let V_t = self.current_svd.v_t.as_ref().expect("Did not calculate U of SVD. This should not happen and indicates a logic error in the library.");
+
+            // we use a functional style calculation here that is more easy to
+            // parallelize with rayon later on. The only disadvantage is that
+            // we don't short circuit anymore if there is an error in calculation,
+            // but since that is the sad path anyways, we don't care about a
+            // performance hit in the sad path.
+            let result: Result<Vec<()>, Model::Error> = jacobian_matrix
+                .par_column_iter_mut()
                 .enumerate()
                 .map(|(k, mut jacobian_col)| {
                     // weighted derivative matrix