+++ CHANGES.md
@@ -3,6 +3,24 @@
 This is the changelog for the `varpro` library.
 See also here for a [version history](https://crates.io/crates/varpro/versions).
 
+## 0.11.0
+
+* Removed `new` and `minimize` associated functions of the `LevMarSolver`
+  type.
+* Require `Send` and `Sync` bounds for `BasisFunction` implementors.
+* Expose parrallel calculations for `LevMarProblem` using extra generic
+  arguments and the rayon dependency. Use the `parallel` feature flag
+  to enable.
+* Require `Send` and `Sync` trait bounds on all base functions for building
+  separable models whether or not the calculations are to be run in parallel.
+  This should not pose restrictions in practice for sane models. If `Send`
+  and `Sync` bounds cannot be satisfied there's always the possibility to
+  implement `SeparableNonlinearModel` by hand.
+
+## 0.10.1
+
+Documentation updates
+
 ## 0.10.0
 
 - Updated dependencies to current versions.